requirements.txt
# Core Dependencies
google-genai>=0.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# API & Web
requests>=2.31.0
beautifulsoup4>=4.12.0

# UI Frameworks
streamlit>=1.28.0
flask>=3.0.0
flask-cors>=4.0.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0

# Storage
sqlalchemy>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Observability
structlog>=23.1.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
